requirements.txt
streamlit
openai
python-dotenv
jinja2
pyyaml
scikit-learn